+++ frontend/src/main/java/com/my/pet/project/mybank/frontend/dto/CashRequest.java
@@ -1,7 +1,9 @@
 package com.my.pet.project.mybank.frontend.dto;
 
+import java.math.BigDecimal;
+
 public record CashRequest(
         Long accountId,
-        int value,
+        BigDecimal value,
         String action
 ) {}
